% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BATgate.R
\name{BATgate}
\alias{BATgate}
\title{BATgate: An algorithm for automated flow cytometry analysis of basophil activation data}
\usage{
BATgate(workingdir, fluorophores, SSCno, medcontroltube)
}
\arguments{
\item{fluorophores}{List of fluorophores, ordered the same as in the fset}

\item{SSCno}{The number of clusters within the SSC distribution}

\item{medcontroltube}{The sample number containing medium only or the control condition in the BAT assay,
used to calculate the MFI for definition of CD63hi basophils}

\item{wkdir}{The location of the working directory, where all of the .fcs files for a BAT experiment are located}
}
\value{
This function allows for data-driven flow cytometry analysis of basophil activation data

A csv file containing basophil flow cytometry output including percentage of CD63hi basophils,
       A pdf file containing graphical gating of flow cytometry analysis for visual inspection
}
\description{
BATgate: An algorithm for automated flow cytometry analysis of basophil activation data
}
\examples{
BATgate(wkdir, c("Time","FSC-A","FSC-H","SSC-A","CD63","CCR3"), 2, 1)
BATgate()
}
\author{
Sarita Patil, \email{sarita.patil@mgh.harvard.edu}
}
\keyword{cluster}
\keyword{models,}
