% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BATmeasures.R
\name{BATmeasures}
\alias{BATmeasures}
\title{BATmeasures: Measures of basophil activation testing}
\usage{
BATmeasures(data, groupvariables, output.file)
}
\arguments{
\item{data}{A dataframe that contains several variables: 
groupvariables which is a grouping variable that identifies all the conditions in a dose-response curve, agconc which is the antigen concentration used in basophil stimulation, and CD63hi which is the percentage of CD63hi basophils from flow analysis in each stimulation condition)}

\item{groupvariables}{The variable name of the grouping variable from the data}

\item{output.file}{The name of the output file}
}
\value{
This function analyzes basophil activation dose-response curve to produce measures of basophil activation

A csv file containing calculated AUC and ED50 values based on input data
}
\description{
BATmeasures: Measures of basophil activation testing
}
\examples{
BATmeasures(data, patientid, output.file="BAT_AUC_ED50.csv")
}
\author{
Sarita Patil, \email{sarita.patil@mgh.harvard.edu}
}
\keyword{models,}
\keyword{nonlinear}
