% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BATgate.R, R/BATmeasures.R
\name{pkgLoad}
\alias{pkgLoad}
\alias{pkgLoad}
\title{BATgate: An algorithm for automated flow cytometry analysis of basophil activation data}
\usage{
pkgLoad(x)

pkgLoad(x)
}
\arguments{
\item{wkdir}{The location of the working directory, where all of the .fcs files for a BAT experiment are located}

\item{fluorophores}{List of fluorophores, ordered the same as in the fset}

\item{SSCno}{The number of clusters within the SSC distribution}

\item{medcontroltube}{The sample number containing medium only or the control condition in the BAT assay,
used to calculate the MFI for definition of CD63hi basophils}

\item{data}{A dataframe that contains several variables: 
groupvariables which is a grouping variable that identifies all the conditions in a dose-response curve, agconc which is the antigen concentration used in basophil stimulation, and CD63hi which is the percentage of CD63hi basophils from flow analysis in each stimulation condition)}

\item{groupvariables}{The variable name of the grouping variable from the data}

\item{output.file}{The name of the output file}
}
\value{
This function allows for data-driven flow cytometry analysis of basophil activation data

A csv file containing basophil flow cytometry output including percentage of CD63hi basophils,
       A pdf file containing graphical gating of flow cytometry analysis for visual inspection

This function analyzes basophil activation dose-response curve to produce measures of basophil activation

A csv file containing calculated AUC and ED50 values based on input data
}
\description{
BATgate: An algorithm for automated flow cytometry analysis of basophil activation data

BATmeasures: Measures of basophil activation testing
}
\examples{
BATgate(wkdir, c("Time","FSC-A","FSC-H","SSC-A","CD63","CCR3"), 2, 1)
BATmeasures(data, patientid, output.file="BAT_AUC_ED50.csv")

}
\author{
Sarita Patil, \email{sarita.patil@mgh.harvard.edu}

Sarita Patil, \email{sarita.patil@mgh.harvard.edu}
}
\keyword{cluster}
\keyword{models,}
\keyword{nonlinear}
